The seed data is complete with the appropriate time system for event. We should consider deleting the :event_time var bc event date includes both date and time

+++ db/migrate/20181205195109_create_events.rb
@@ -2,7 +2,7 @@ class CreateEvents < ActiveRecord::Migration[5.0]
   def change
     create_table :events do |e|
      e.string :name
-     e.float :price
+     e.integer :price
      e.datetime :event_date
      e.datetime :event_time
      e.boolean :alcohol
